Data is now loaded off of seperate scripts
As it turns out, linking an assembly that Unity can't see, does not
allow you to correctly link objects to the scripts in the unity
interface on Monobehaviours where unity can't find its dependancies.
You can still use those scripts though.  Finally made the required
changes so that it becomes possible to build proper release builds of
the assembly and have them work.