Turn the key changes.
Turn the Key now solves instantly if a valid turn command is supplied,
that key is the one that should be turned first, and there are no other
modules besides turn the key left to solve on that bomb.  The
consequence of this is that the timer will be instantly dropped to match
the correct key turn time, and not restored to what it was.