note about cruel code
The author has explicitly requested that there is no TP support for the
module.  Will be difficult to enforce, unless he removes the TP support
code module side.  Until the author rescinds his decision, no shim or
explicit support for Cruel code is to be added (if removed module side.)