Fix module issues with !enablecamerawall
Instead of hiding the bomb by moving it out of view,  instead, turned
off ALL of the main camera culling mask flags, (after saving a copy of
course.)  As soon as bomb ends, or when !disablecamerawall in issued,
the culling mask is restored.

This does mean the screen will be mostly black except for where camera
views exist.