Fix a bug where !take-ers weren’t told when the module was assigned to them.
The extent of this change requires some explanation. The claim/unclaim logic was all over the place. It wasn’t clear which parts of the code checked which conditions (claim limit, cooldown, moderator powers, etc.) and many were checked in multiple places. The bug was a result of this complexity.
To address this, this is now streamlined: there’s a TryClaim() (similar to earlier CanClaimNow()) that will check all conditions and either claim or add the player to the claim queue. Instead of a Tuple<bool, string>, we now use a ClaimResult class.
Then there is SetClaimedBy(), which directly sets the module as claimed, no questions asked.
Similarly, setUnclaimed() will directly mark the module as unclaimed. All callers have been checked to ensure that all conditions are checked (which isn’t many: it must be claimed, and either by you or you’re a moderator).
We also have the claim precedence logic in one place: ProcessClaimQueue() (formerly AutoAssignModule()). All other code (including the takeover coroutine) now only needs to unclaim modules and this will correctly assign modules to users, giving precedence to takeover attempts and using TryClaim() to check all claim conditions.