Fix Shim Command Processors
The optimization you asked me to test for the Shims was apparently not good as it broke all commands for the modules. I swore I tested one of them and it worked but I guess I messed up somewhere. Whatever the case I changed all of them back to what I had them before as they work like this.